TASK: Replace remanding usages from odd `NodeAggregate::getNodes()` and ensure StructureAdjustments are deterministically sorted
Extract `requireOrderedOriginDimensionSpacePoints` into own trait for sharing
